feat(share): score labels are the expected level, not the argmax
JEV answers a score question with a distribution; storing its argmax
collapsed 47 people into "warmth 3" who actually spread from 2.5 to
3.97. The saved label is now the expected level (Σ level × p — the same
number as JEV's own `score`), so the tiers order by how sure JEV was.
Rewritten for every saved person from the request cache at no cost.

+++ packs/ingestion/primitives/share/models.py
@@ -108,7 +108,8 @@ class JevLabels:
 
     choices: dict[str, str]
     choice_p: dict[str, float]
-    scores: dict[str, int]
+    # Expected level (0.00-4.00 for warmth), not the argmax: the tiers spread.
+    scores: dict[str, float]
     probabilities: dict[str, float]
